Implement sleep_until for Timestamp
- Remove `sleep_until(timestamp: Timestamp)` in `EtherHarnessParams`
  and `Erc20HarnessParams`, and reintroduce it in `Timestamp`'s `impl`
  block so that the function can be reused for Bitcoin, Ether and
  ERC20 tests.
