+++ application/comit_node/src/swap_protocols/rfc003/timestamp.rs
@@ -1,4 +1,7 @@
-use std::time::SystemTime;
+use std::{
+    thread::sleep,
+    time::{Duration, SystemTime},
+};
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Deserialize, Serialize)]
 #[serde(transparent)]
@@ -18,7 +21,14 @@ impl Timestamp {
         Self(self.0 + seconds)
     }
 
-    pub fn diff(self, other: Timestamp) -> u32 {
+    pub fn sleep_until(timestamp: Timestamp) {
+        let duration = timestamp.diff(Timestamp::now());
+        let buffer = 2;
+
+        sleep(Duration::from_secs((duration + buffer).into()));
+    }
+
+    fn diff(self, other: Timestamp) -> u32 {
         self.0.checked_sub(other.0).unwrap_or(0)
     }
 }
