+++ application/comit_node/tests/htlc_harness/ether_harness.rs
@@ -8,7 +8,7 @@ use ethereum_support::{
     web3::{transports::EventLoopHandle, types::Address},
     EtherQuantity,
 };
-use std::{sync::Arc, time::Duration};
+use std::sync::Arc;
 use tc_web3_client;
 use testcontainers::{images::parity_parity::ParityEthereum, Container, Docker};
 
@@ -37,13 +37,6 @@ impl EtherHarnessParams {
             htlc_secret_hash: secret_hash,
             ..self
         }
-    }
-
-    pub fn sleep_until(timestamp: Timestamp) {
-        let duration = timestamp.diff(Timestamp::now());
-        let buffer = 2;
-
-        ::std::thread::sleep(Duration::from_secs((duration + buffer).into()));
     }
 }
