+++ application/comit_node/tests/rfc003_erc20_htlc.rs
@@ -14,6 +14,7 @@ use crate::{
     ethereum_wallet::transaction::UnsignedTransaction,
     htlc_harness::{erc20_harness, CustomSizeSecret, Erc20HarnessParams, SECRET},
 };
+use comit_node::swap_protocols::rfc003::Timestamp;
 use ethereum_support::{Bytes, H256, U256};
 use spectral::prelude::*;
 use testcontainers::clients::Cli;
@@ -115,7 +116,7 @@ fn given_deployed_erc20_htlc_when_refunded_after_expiry_time_then_tokens_are_ref
     assert_eq!(client.token_balance_of(token, alice), U256::from(600));
 
     // Wait for the contract to expire
-    Erc20HarnessParams::sleep_until(harness_params.htlc_refund_timestamp);
+    Timestamp::sleep_until(harness_params.htlc_refund_timestamp);
     client.send_data(htlc_address, None);
 
     assert_eq!(client.token_balance_of(token, htlc_address), U256::from(0));
@@ -249,7 +250,7 @@ fn given_htlc_and_refund_should_emit_refund_log_msg() {
     });
 
     // Wait for the contract to expire
-    Erc20HarnessParams::sleep_until(harness_params.htlc_refund_timestamp);
+    Timestamp::sleep_until(harness_params.htlc_refund_timestamp);
     // Send correct secret to contract
     let transaction_receipt = client.send_data(htlc_address, None);
