+++ application/comit_node/tests/rfc003_ether_htlc.rs
@@ -11,6 +11,7 @@ pub mod htlc_harness;
 pub mod parity_client;
 
 use crate::htlc_harness::{ether_harness, CustomSizeSecret, EtherHarnessParams, SECRET};
+use comit_node::swap_protocols::rfc003::Timestamp;
 use ethereum_support::{Bytes, EtherQuantity, H256};
 use spectral::prelude::*;
 use testcontainers::clients::Cli;
@@ -66,7 +67,7 @@ fn given_deployed_htlc_when_refunded_after_expiry_time_then_money_is_refunded() 
     );
 
     // Wait for the contract to expire
-    EtherHarnessParams::sleep_until(harness_params.htlc_refund_timestamp);
+    Timestamp::sleep_until(harness_params.htlc_refund_timestamp);
     client.send_data(htlc, None);
 
     assert_eq!(
@@ -134,7 +135,7 @@ fn given_htlc_and_refund_should_emit_refund_log_msg() {
         ether_harness(&docker, harness_params.clone());
 
     // Wait for the timelock to expire
-    EtherHarnessParams::sleep_until(harness_params.htlc_refund_timestamp);
+    Timestamp::sleep_until(harness_params.htlc_refund_timestamp);
     let transaction_receipt = client.send_data(htlc, None);
 
     assert_that(&transaction_receipt.logs).has_length(1);
